+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# upcoming
+
+* BREAKING CHANGE: Update `drf-form-field` template to match https://github.com/incuna/style-guide/blob/137a7c092703292ec93713088e10cdce650b5a6a/templates/components/forms.html
+
 ## 5.1.0
 
 * Add `drf-form-field` directive and `drfFieldErrors` service